+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0
-- Skip flutter pub get for example projects
+- Determine if a project uses dart or flutter automatically
+- Skip flutter pub get for flutter example projects
 
 ## 1.0.0
 - Initial version